Simplify distribution of version.py
Close #2018
